An important commit. At the end `The Game` is runnable and works good :).

+++ src/tables.rs
@@ -1,16 +1,11 @@
 use hero;
 
-const MaxPlayerLevel: u32 = 8;
+const MaxPlayerLevel: usize = 8;
 
-pub const ExpLevel_Table: [u32; MaxPlayerLevel as usize] = [
-    10, 20, 50, 100, 250, 500, 1000, 3000
-];
+pub const ExpLevel_Table: [usize; MaxPlayerLevel as usize] =
+    [10, 20, 50, 100, 250, 500, 1000, 3000];
 
-pub const HPLevel_Table: [u32; MaxPlayerLevel as usize] = [
-    10, 20, 30, 50, 80, 130, 210, 340
-];
+pub const HPLevel_Table: [usize; MaxPlayerLevel as usize] = [10, 20, 30, 50, 80, 130, 210, 340];
 
 // percents
-pub const BaseSkill_Table: [u32; hero::MaxSkills as usize] = [
-    30, 20
-];
+pub const BaseSkill_Table: [usize; hero::MaxSkills as usize] = [30, 20];
